Began updating tests
