+++ packages/apollos-data-connector-rock/src/people/data-source.js
@@ -32,33 +32,6 @@ export default class Person extends RockApolloDataSource {
     return null;
   };
 
-  // Gets a collection of all dataviews a user is in
-  // Returns an array of dataview guids
-  getPersonas = async ({ categoryId }) => {
-    const {
-      dataSources: { RockConstants, Auth },
-    } = this.context;
-
-    // Get current user
-    const { id } = await Auth.getCurrentPerson();
-
-    // Get the entity type ID of the Person model
-    const personEntityTypeId = await RockConstants.modelType('Person');
-
-    // Rely on custom code without the plugin.
-    // Use plugin, if the user has set USE_PLUGIN to true.
-    // In general, you should ALWAYS use the plugin if possible.
-    const endpoint = get(ApollosConfig, 'ROCK.USE_PLUGIN', false)
-      ? 'Apollos/GetPersistedDataViewsForEntity'
-      : 'DataViews/GetPersistedDataViewsForEntity';
-
-    // Return a list of all dataviews by GUID a user is a memeber
-    return this.request(endpoint)
-      .find(`${personEntityTypeId.id}/${id}?categoryId=${categoryId}`)
-      .select('Guid')
-      .get();
-  };
-
   mapGender = ({ gender }) => {
     // If the gender is coming from Rock (an int) map into the string value.
     if (typeof gender === 'number') {
